requirements.txt
streamlit
numpy
pillow
pybase64
opencv-python
tensorflow==2.0
scikit-learn
Keras

